On representation change, force the re-download of init if present
On shaka packager, the separate init component is needed to update
the esinfo for the given ES stream, so on every change of stream
we need to get the init information into the stream ahead of the
reconfigured video.

This change marks the adaptive demux stream as need_header=TRUE
on every representation change.